+++ concordium-consensus/package.yaml
@@ -501,4 +501,3 @@ benchmarks:
     - temporary >= 1.3
     - criterion
     - deepseq
-    - HUnit >= 1.6
